+++ src/main/java/org/c4marathon/assignment/domain/account/controller/AccountController.java
@@ -1,9 +1,9 @@
 package org.c4marathon.assignment.domain.account.controller;
 
-import org.c4marathon.assignment.domain.account.dto.CreateResponseDto;
-import org.c4marathon.assignment.domain.account.dto.RemittanceRequestDto;
-import org.c4marathon.assignment.domain.account.dto.RemittanceResponseDto;
-import org.c4marathon.assignment.domain.account.dto.SavingRequestDto;
+import org.c4marathon.assignment.domain.account.dto.response.CreateResponseDto;
+import org.c4marathon.assignment.domain.account.dto.request.RemittanceRequestDto;
+import org.c4marathon.assignment.domain.account.dto.response.RemittanceResponseDto;
+import org.c4marathon.assignment.domain.account.dto.request.SavingRequestDto;
 import org.c4marathon.assignment.domain.account.service.AccountService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
