--- src/main/java/org/c4marathon/assignment/domain/account/dto/RemittanceResponseDto.java
@@ -1,5 +1,0 @@
-package org.c4marathon.assignment.domain.account.dto;
-
-
-public record RemittanceResponseDto(String responseMsg) {
-}
