+++ src/main/java/org/c4marathon/assignment/domain/account/dto/request/RemittanceRequestDto.java
@@ -1,4 +1,4 @@
-package org.c4marathon.assignment.domain.account.dto;
+package org.c4marathon.assignment.domain.account.dto.request;
 
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
