+++ src/main/java/org/c4marathon/assignment/domain/account/dto/response/CreateResponseDto.java
@@ -1,0 +1,5 @@
+package org.c4marathon.assignment.domain.account.dto.response;
+
+
+public record CreateResponseDto(String responseMsg) {
+}
