+++ src/main/java/org/c4marathon/assignment/domain/account/dto/response/RemittanceResponseDto.java
@@ -1,0 +1,5 @@
+package org.c4marathon.assignment.domain.account.dto.response;
+
+
+public record RemittanceResponseDto(String responseMsg) {
+}
